+++ README.md
@@ -1,14 +1,47 @@
-# TODO Test Task for Israel IT
-
 <p align="center">
-  <a href="https://github.com/mazurdev/angular-full-starter">
-    <img src="https://raw.githubusercontent.com/mazurdev/angular-full-starter/master/meta-assets/angular-logo.png" alt="Angular Logo" width="96" height="96">
+  <a href="https://github.com/mazurdev/todo-test">
+    <img src="https://raw.githubusercontent.com/mazurdev/todo-test/master/meta-assets/preview-logo-angular.png" alt="Angular Logo" width="102" height="127">
   </a>
-  <h1 align="center">ANGULAR FULL STARTER</h1>
-  <p align="center">Angular starter, which combines all the necessary components for writing a complete project</p>
-  
+  <h1 align="center">TODO Test Task for Israel IT</h1>
 </p>
 
 <p align="center">
-  <img src="https://raw.githubusercontent.com/mazurdev/angular-full-starter/master/meta-assets/preview.gif">
+  <img src="https://raw.githubusercontent.com/mazurdev/todo-test/master/meta-assets/preview.gif">
 </p>
+  
+## Table of Content
+
+  * [Getting Started](#getting-started)
+  * [Stack](#stack)
+  * [Useful Commands](#useful-commands)
+  * [License](#license)
+
+## Getting started
+```bash
+1 - git clone https://github.com/mazurdev/todo-test.git
+2 - Go to "todo-test" folder
+3 - npm install
+4 - Change the name of the project "todo-test" to your in files: "angular.json", "package.json", "README.md"
+5 - Run server: npm run server
+6 - Run Client: npm run dev
+```
+
+## Stack
+
+  * Angular 9+
+  * JSON Server
+  * Angular Material
+  * RxJs
+
+## Useful Commands
+
+  * `npm run dev` - starts a dev server
+  * `npm run server` - starts a json server
+  * `npm run start` - it is startup-script for Heroku Deploy
+  * `npm run build` - production build 
+  * `npm run analyze:bundle` - analyze bundle application using webpack-bundle-analyzer
+  * `npm run postinstall` - run ngcc compile libraries
+
+## License
+
+[MIT License](https://raw.githubusercontent.com/mazurdev/todo-test/master/LICENSE)
