+++ proxy.conf.json
@@ -1,6 +1,6 @@
 {
   "/api": {
-    "target": "http://localhost:3000",
+    "target": "http://localhost:4000",
     "secure": false,
     "changeOrigin": true
   }
